🎨Feat:"query 구조 개선 모듈화"
Resolves: #19
